// frontend/src/app/client/models/Message.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Message = {
    id: string;
    timestamp: string;
    collectionId?: (string | null);
    collectionName?: (string | null);
    topic: string;
    message: string;
};